+++ crates/uv/src/commands/pip_install.rs
@@ -24,8 +24,7 @@ use uv_client::{Connectivity, FlatIndex, FlatIndexClient, RegistryClient, Regist
 use uv_dispatch::BuildDispatch;
 use uv_fs::Simplified;
 use uv_installer::{
-    compile_tree, BuiltEditable, Downloader, NoBinary, Plan, Planner, Reinstall, ResolvedEditable,
-    SitePackages,
+    BuiltEditable, Downloader, NoBinary, Plan, Planner, Reinstall, ResolvedEditable, SitePackages,
 };
 use uv_interpreter::{Interpreter, PythonEnvironment};
 use uv_normalize::PackageName;
@@ -36,7 +35,7 @@ use uv_resolver::{
 use uv_traits::{ConfigSettings, InFlight, NoBuild, SetupPyStrategy};
 
 use crate::commands::reporters::{DownloadReporter, InstallReporter, ResolverReporter};
-use crate::commands::{elapsed, ChangeEvent, ChangeEventKind, ExitStatus};
+use crate::commands::{compile_and_report, elapsed, ChangeEvent, ChangeEventKind, ExitStatus};
 use crate::printer::Printer;
 use crate::requirements::{ExtrasSpecification, RequirementsSource, RequirementsSpecification};
 
@@ -684,26 +683,7 @@ async fn install(
     }
 
     if compile {
-        let start = std::time::Instant::now();
-        let files = compile_tree(venv.site_packages(), venv.python_executable())
-            .await
-            .with_context(|| {
-                format!(
-                    "Failed to bytecode compile {}",
-                    venv.site_packages().display()
-                )
-            })?;
-        let s = if files == 1 { "" } else { "s" };
-        writeln!(
-            printer,
-            "{}",
-            format!(
-                "Bytecode compiled {} in {}",
-                format!("{files} file{s}").bold(),
-                elapsed(start.elapsed())
-            )
-            .dimmed()
-        )?;
+        compile_and_report(&mut printer, &venv).await?;
     }
 
     // TODO(konstin): Also check the cache whether any cached or installed dist is already known to
